Extract Gateway into Gateway.Connection and Gateway.Broker
Using GenStage, we can provide better resiliency than GenEvent to dispatch
incoming Gateway events to consumers.

Gateway.Connection will own the Websocket and handle communication with
the Gateway. Once events are received and parsed, they are dispatched to
the Broker (a GenStage producer) which then broadcasts them to GenStage
consumers.

The client can thus create as many GenStage consumers as desired and
subscribe them to the Broker.

If the Connection process crashes ( due to network issues for example), the
Broker still operates and may broadcasts buffered events. If a consumer process
crashes, the Connection and other consumers are not impacted.

For now we do not have any mechanism providing back-pressure or
load-shedding (as the Broker buffer is unbounded). We could limit the size of
the Broker buffer and reject new events, blocking the Connection read loop
(back-pressure) or we could remove old events from the queue to make room for
new ones (load-shedding).